\name{blighty-internal}
\title{blighty internal function and datasets}
\alias{sqlimits}
\alias{Coastline.of.England}
\alias{Coastline.of.Scotland}
\alias{Coastline.of.Wales}
\alias{England.Lake.Conniston}
\alias{England.Lake.Crummock.Water.Buttermere}
\alias{England.Lake.Derwent.Reservoir}
\alias{England.Lake.Derwent.Water}
\alias{England.Lake.Rutland.Water}
\alias{England.Lake.Ullswater}
\alias{England.Lake.Windermere}
\alias{England.River.Nene}
\alias{England.River.Ouse}
\alias{England.River.Severn}
\alias{England.River.Thames}
\alias{England.River.Trent}
\alias{Hebrides.Arran}
\alias{Hebrides.Barra}
\alias{Hebrides.Benbecular}
\alias{Hebrides.Coll}
\alias{Hebrides.Islay}
\alias{Hebrides.Jura}
\alias{Hebrides.Lewis}
\alias{Hebrides.Mull}
\alias{Hebrides.North.Uist}
\alias{Hebrides.Rum}
\alias{Hebrides.Skye}
\alias{Hebrides.South.Uist}
\alias{Hebrides.Tiree}
\alias{Isle.of.Anglesea}
\alias{Isle.of.Lundy}
\alias{Isle.of.Man}
\alias{Isle.of.Wight}
\alias{Orkneys.Eday}
\alias{Orkneys.Fair.Isle}
\alias{Orkneys.Hoy}
\alias{Orkneys.Mainland}
\alias{Orkneys.North.Ronaldsay}
\alias{Orkneys.Rousay}
\alias{Orkneys.Sanday}
\alias{Orkneys.Shapinsay}
\alias{Orkneys.South.Ronaldsay}
\alias{Orkneys.Stronsay}
\alias{Orkneys.Westray}
\alias{Scillies.Annet}
\alias{Scillies.Bryher}
\alias{Scillies.Eastern.Islands}
\alias{Scillies.Gugh}
\alias{Scillies.Round.Island}
\alias{Scillies.Samson}
\alias{Scillies.St.Agnes}
\alias{Scillies.St.Martins}
\alias{Scillies.St.Marys}
\alias{Scillies.Tean}
\alias{Scillies.Tresco}
\alias{Scillies.White.Island}
\alias{Scotland.Loch.Arkaig}
\alias{Scotland.Loch.Awe}
\alias{Scotland.Loch.Cluanie}
\alias{Scotland.Loch.Ericht}
\alias{Scotland.Loch.Lochy}
\alias{Scotland.Loch.Lomond}
\alias{Scotland.Loch.Maree}
\alias{Scotland.Loch.Monar}
\alias{Scotland.Loch.Morar}
\alias{Scotland.Loch.Mullardoch}
\alias{Scotland.Loch.Ness}
\alias{Scotland.Loch.Quoich}
\alias{Scotland.Loch.Rannoch}
\alias{Scotland.Loch.Shin}
\alias{Scotland.Loch.Tay}
\alias{Scotland.River.Clyde}
\alias{Scotland.River.Spey}
\alias{Scotland.River.Tweed}
\alias{Shetlands.Fetlar}
\alias{Shetlands.Foula}
\alias{Shetlands.Mainland}
\alias{Shetlands.Unst}
\alias{Shetlands.Yell}
\alias{Wales.Lake.Claerwen.Craug.Gock.Reservoirs}
\alias{Wales.Lake.Llyn.Brianne.Reservoir}
\alias{Wales.River.Usk}
\alias{set.all}
\alias{set.England}
\alias{set.Scotland}
\alias{set.Scotland.all}
\alias{set.UK}
\alias{set.Wales}
\description{These are internal functions and should not be used directly, although it won't do any harm were one to do so.}
\author{David Lucy}
\keyword{internal}
