\name{blighty-internal}
\title{blighty internal function and datasets}
\alias{sqlimits}
\alias{getobjectnames}
\alias{Border.of.England.and.Scotland}
\alias{Border.of.England.and.Wales}
\alias{Coastline.of.Cornwall}
\alias{Coastline.of.East.Anglia}
\alias{Coastline.of.Eastern.Scotland}
\alias{Coastline.of.North.East.England}
\alias{Coastline.of.North.West.England}
\alias{Coastline.of.Northern.Scotland}
\alias{Coastline.of.South.England}
\alias{Coastline.of.Wales}
\alias{Coastline.of.Western.Scotland}
\alias{Isle.of.Anglesea}
\alias{Isle.of.Arran}
\alias{Isle.of.Benbecular}
\alias{Isle.of.Coll}
\alias{Isle.of.Colonsay}
\alias{Isle.of.Islay}
\alias{Isle.of.Jura}
\alias{Isle.of.Lewis}
\alias{Isle.of.Man}
\alias{Isle.of.Mull}
\alias{Isle.of.North.Uist}
\alias{Isle.of.Rum}
\alias{Isle.of.Skye}
\alias{Isle.of.South.Uist}
\alias{Isle.of.Tiree}
\alias{Isle.of.Wight}
\description{These are internal functions and should not be used directly, although it won't do any harm were one to do so.}
\author{David Lucy}
\keyword{misc}
